+++ src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/components/map-area-select-form-control/map-area-select-form-control.model.ts
@@ -1,5 +1,3 @@
-import { InjectionToken } from '@angular/core';
-
 export interface MapAreaSelectFormControlVM {
     areaValueKey: string;
     addButtonTextKey: string;
@@ -7,21 +5,3 @@ export interface MapAreaSelectFormControlVM {
     deleteButtonTextKey: string;
     coordinatesLabelKey: string;
 }
-
-export const POLYGON_OPTIONS = new InjectionToken<google.maps.PolygonOptions>(
-    'Injection Token for Polygon options'
-);
-
-export const provideMockPolygonOptions = () => ({
-    provide: POLYGON_OPTIONS,
-    useValue: {},
-});
-export const INFO_WINDOW_OPTIONS =
-    new InjectionToken<google.maps.InfoWindowOptions>(
-        'Injection Token for Info Window options'
-    );
-
-export const provideMockInfoWindowOptions = () => ({
-    provide: INFO_WINDOW_OPTIONS,
-    useValue: {},
-});
