+++ src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/map-form-control.model.ts
@@ -1,7 +1,6 @@
 import { MapSearchInputFormControlVM } from 'src/app/pages/orders-new-page/components/areas-data-form-control/components/area-data-dialog/components/map-form-control/components/map-search-input-form-control/map-search-input-form-control.model';
 import { MapAreaSelectFormControlVM } from './components/map-area-select-form-control/map-area-select-form-control.model';
 import { MapPointSelectFormControlVM } from './components/map-point-select-form-control/map-point-select-form-control.model';
-import { InjectionToken } from '@angular/core';
 import { Coordinates } from '@stores/location/location.model';
 
 export interface MapFormControlVM {
@@ -9,23 +8,6 @@ export interface MapFormControlVM {
     mapAreaSelectFormControlVM: MapAreaSelectFormControlVM;
     mapPointSelectFormControlVM: MapPointSelectFormControlVM;
     mapSearchInputFormControlVM: MapSearchInputFormControlVM;
-}
-
-export const MAP_OPTIONS = new InjectionToken<MapOptions>(
-    'Injection Token for Map options'
-);
-
-export const provideMockMapOptions = () => ({
-    provide: MAP_OPTIONS,
-    useValue: {},
-});
-export interface MapOptions {
-    mapTypeId: 'roadmap';
-    disableDoubleClickZoom: boolean;
-    clickableIcons: boolean;
-    mapId: 'DEMO_MAP_ID';
-    zoom: number;
-    streetViewControl: boolean;
 }
 
 export const mapStringToCoordinates = (input: string): Coordinates[] => {
